+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:background="@drawable/login_background_img"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.LoginActivity">
+    tools:context=".view.LoginActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
